Make it easier to reset static state in shadows.
Expand @Implements annotation with a "hasStaticState" value. Before and
after a test is run, we look for any shadows (both custom shadows and
default shadows) that has this value set and call the reset method.

Fixes #1072.
