+++ src/main/java/org/robolectric/internal/ParallelUniverseInterface.java
@@ -8,7 +8,7 @@ import org.robolectric.annotation.Config;
 import org.robolectric.res.ResourceLoader;
 
 public interface ParallelUniverseInterface {
-  public void resetStaticState();
+  void resetStaticState(Config config);
 
   void setUpApplicationState(Method method, TestLifecycle testLifecycle, boolean strictI18n, ResourceLoader systemResourceLoader, AndroidManifest appManifest, Config config);
