+++ src/main/java/org/robolectric/shadows/ShadowContentResolver.java
@@ -34,7 +34,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 
-@Implements(ContentResolver.class)
+@Implements(value = ContentResolver.class, resetStaticState = true)
 public class ShadowContentResolver {
   private int nextDatabaseIdForInserts;
   private int nextDatabaseIdForUpdates;
