+++ src/main/java/org/robolectric/shadows/ShadowProcess.java
@@ -6,7 +6,7 @@ import org.robolectric.annotation.Implements;
 /**
  * Shadows the {@code android.os.Process} class.
  */
-@Implements(android.os.Process.class)
+@Implements(value = android.os.Process.class, resetStaticState = true)
 public class ShadowProcess {
   private static int pid;
